The following changes were merged:
commit 1b88ecde3f4de05c56ec7c075b0188ca0a89337f
Author: Max Sagebaum <sagebaum@mathcces.rwth-aachen.de>
Date:   Wed Apr 3 10:55:33 2013 +0200

    Allow adouble objects to be initialized late

    This functionality is specially required for large codes that allocate
    memory using malloc/calloc and we only wish to substitute the double
    datatype with adouble for tracing purposes. With malloc/calloc the
    adouble constructor is not called, creating problems later in the code.
    Late initialization defers some of the initialization to the first use
    of an allocated space as an adouble object. This is only possible because
    the class heirarchy in ADOL-C does not require polymorphism, and virtual
    tables to be allocated, which can only be done in a chained superclass
    constructor call.

    The functionality is disabled by default and can be enabled by a configure
    switch --enable-lateinit

    This code is experimental and has only be tested with particular codes.

    From: Max Sagebaum <sagebaum@mathcces.rwth-aachen.de>
    CC: Andrea Walther <andrea.walther@uni-paderborn.de>
    CC: Nicolas Gauger <gauger@mathcces.rwth-aachen.de>

commit 1a553a5e24b575e955985db19b01b6726ba8b6bf
Author: Kshitij Kulshreshtha <kshitij@math.upb.de>
Date:   Tue Apr 2 11:14:50 2013 +0200

    add dvi file to gitignore


commit 2b75a6a7802b141092fbfdc948f735617d713c8e
Author: Kshitij Kulshreshtha <kshitij@math.upb.de>
Date:   Tue Apr 2 11:14:37 2013 +0200

    remove as many compiler warnings from gcc -Wall as possible

    only -Wunused-value remains, as we do indeed ignore some
    values computed when we are in INDO, NONL_IND or
    INT_FOR, INT_REV modes.

+++ ADOL-C/src/tape_handling.cpp
@@ -114,7 +114,7 @@ StoreManagerLocint::StoreManagerLocint(const StoreManagerLocint *const stm,
 #endif
     head = stm->head;
     indexFree = new locint[maxsize];
-    for (int i = 0; i < maxsize; i++)
+    for (size_t i = 0; i < maxsize; i++)
 	indexFree[i] = stm->indexFree[i];
 }
